xmonad: Dynamically determine default browser to open

+++ xmonad/install.sh
@@ -1,10 +1,14 @@
-#!/bin/sh
+#!/bin/bash
 
 SRCFLD=$(dirname $(readlink -m "$0"))
 TGTFLD="$HOME/.xmonad"
 
-ln -sfv "${SRCFLD}" ${TGTFLD}
+if [ ! -e "${TGTFLD}" ]; then
+    ln -sfv "${SRCFLD}" ${TGTFLD}
+fi
+
+TGTFLD="${HOME}/.local/bin"
+mkdir -p "${TGTFLD}"
+find "${SRCFLD}/utils" -type f -executable -print0 | xargs -0 ln -sfvt "${TGTFLD}"
 
 pydemx $SRCFLD/xmobar.pydemx
-
-
